# fixes bug where labelmap said there was no imagery